Add support for installing servers outside of ST "catalog"
A large number of files within directories that ST considers part of the
"catalog" make ST less performant, making catalog queries and ST startup
take a longer time.

Introduced a class method that plugins can override to request being
installed in the "Data/LSP-servers/[PackageName]" directory which is
part of the user settings but not inside directories that are part of
the "catalog".

This need to be done in an opt-in way instead of just switching directly
in lsp_utils as we need to allow the update flow to run for dependent
packages so that the "plugin_unload" event clears out the old directory.